+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         tools:targetApi="34"
         android:enableOnBackInvokedCallback="true">
         <activity
-            android:name=".view.MainActivity"
+            android:name=".view.activities.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.YouRun" >
             <intent-filter>
@@ -24,7 +24,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".view.AppExpActivity"
+            android:name=".view.activities.AppExpActivity"
             android:exported="true"
             android:theme="@style/Theme.YouRun" />
