+++ app/src/main/java/com/example/yourun/view/activities/AppExpActivity.kt
@@ -1,8 +1,9 @@
-package com.example.yourun.view
+package com.example.yourun.view.activities
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.yourun.R
+import com.example.yourun.view.fragments.AppExpFragment
 
 class AppExpActivity : AppCompatActivity() {
