+++ app/src/main/java/com/example/yourun/view/fragments/AppExpFragment4.kt
@@ -1,4 +1,4 @@
-package com.example.yourun.view
+package com.example.yourun.view.fragments
 
 import android.os.Bundle
 import android.text.SpannableString
@@ -15,20 +15,20 @@ import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import com.example.yourun.R
 
-class AppExpFragment2 : Fragment() {
+class AppExpFragment4: Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_app_exp2, container, false)
+        return inflater.inflate(R.layout.fragment_app_exp4, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
         view.findViewById<Button>(R.id.next_btn).setOnClickListener {
-            val nextFragment = AppExpFragment3()
+            val nextFragment = AppExpFragment5()
             parentFragmentManager.beginTransaction()
                 .replace(R.id.app_exp_fragment_container, nextFragment)
                 .addToBackStack(null)
@@ -48,10 +48,10 @@ class AppExpFragment2 : Fragment() {
             }
         }
 
-        val txt_what_character = view.findViewById<TextView>(R.id.txt_app_exp_what_character)
+        val txt_who_run = view.findViewById<TextView>(R.id.txt_app_exp_who_run)
 
-        val fullText = txt_what_character.text.toString()
-        val targetText = "캐릭터"
+        val fullText = txt_who_run.text.toString()
+        val targetText = "더 많이, 더 꾸준히"
         val spannableString = SpannableString(fullText)
         val startIndex = fullText.indexOf(targetText)
         val endIndex = startIndex + targetText.length
@@ -64,7 +64,7 @@ class AppExpFragment2 : Fragment() {
                 SPAN_EXCLUSIVE_EXCLUSIVE
             )
         }
-        txt_what_character.text = spannableString
+        txt_who_run.text = spannableString
 
     }
 }
